--- a/src/bp/nlu-core/intents/svm-intent-classifier.ts
+++ b/src/bp/nlu-core/intents/svm-intent-classifier.ts
@@ -65,6 +65,8 @@ export class SvmIntentClassifier implements IntentClassifier {
       list_entities,
       pattern_entities
     }
+
+    progress(1)
   }
 
   serialize(): string {