Allright, last bulky PR before I start focussing on something else.

### This PR

- isolate ctx classification in class
- intent and ctx share the same svm classifier (In this commit, I also remove any notion of none intent in in all of the pipeline expect the `oos-intent-clf`.)
- make the slot tagger look like the intent classifier a bit more (similar interface) + have one slot tagger per intent

This last point changes results when there is more than one intent. I didnt noticed this using bitfan alone as it test slots for one dummy intent.

To say the least, this last point deserve some attention and investigating. I maybe should have dedicated a single PR to this last bullet point, but what is coded is coded.

So,

I went back using the old nlu-testing module with this [bot](https://github.com/botpress/botpress/tree/master/modules/nlu-testing/src/bot-templates/bp-nlu-slot-extraction).

Those are the results I've got:

**Accuracy (%)**
| seed | master | composite_pipeline | this branch |
|------|--------|--------------------|-------------|
| 42   | 55.1   | 51.7               | 57.3        |
| 69   | 55.1   | 53.9               | 60.7        |
| 666  | 55.1   | 53.9               | 60.7        |
| 1234 | 53.9   | 53.9               | 60.7        |

**\*remember that one test flipping ~1.1%\***

**Train Speed (s)**

| master | composite_pipeline | this branch |
|--------|--------------------|-------------|
| ~6s    | ~6s                | ~2s         |

It seems like this new approach of one slot tagger per intent is not only faster to train, but makes better predictions...

### Reasons to merge this

1 - Context classifier uses the exact same code as the intent classifier. I see this as a victory.
2 - Slot tagger seems to perform quite well.
3 - Even in a scenario where the slot tagger would have been a bit weaker, this still should have been merged. I really don't want slot results to vary because I have less or more intents. Other intents should not impact the result of slot tagging for one intent. This goes in the same line as all we did so fare in the recent NLU modifications, that is to say : try to make nlu components the most indpendant possible.

### Next steps

- Address all comments I've received so far + notes and TODO I've accumulated and try to make something mergeable.
- Get rid of notion of context classification. Replace by multi-level intent or sub intents.  global slots can be adressed here without problem. What do you think about this: a global slot is "a slot for context". Anyway, more on this in another PR.
- Extract whats comon between all componants and take a step in writing an sdk for custom pipeline components. 

@EFF let's discuss which one of the above should be my next step.