@franklevasseur 2021-01-29
Allright I have tested few things and updated the PR's description.

This PR is ready for review now.